add support for pulling in release zip code instead of cloning the tags
The remote-data-blocks integration uses a built version of the code which is available as a zip file for each release. A new field `releaseZipFileName` is added to the config which if present will indicate that the release zip file with that name should be downloaded and extracted into the folder for the integration.