refactor version handling to only use prefix for git operations
The existing implementation only stripped the prefix from the version for comparison and everywhere else it used the full version including the prefix. As the folder structure shouldn't have the version prefix, it seemed easier to flip the implementation we always use the version without the prefix including in the config.json and only add the prefix when checking if the tag exists in git and pulling the tag code.